refactoring typechecker
Instead of a huge if-elif-else case with lots of is_kind_of
questions there is only one is_kind_of left that does the job.
This works because in many cases the type checker now checks
for w_numeric instead of w_float, thus accepting a Fixnum where
it didn't previously.