Added an alternative to check_type_errors
A different approach to the check_type_errors function is one that is less
generic and focuses on the needs of the Math modules. It's called converter for
now because it actually just converts the input variables to float (but in a
stricter way as Coerce does).
ldexp still depends on check_type_errors.