Changed test_type_error
This test was not entirely correct:
It expected the raised error of a call such as
"function_taking_two_floats('Bang!')"
to be a TypeError since the first argument was a string. The truth is that (at
least in the math module of ruby 1.9.3) the raised error should be an
ArgumentError because the function is called with one instead of two arguments.
In other words: Checking for an ArgumentError should be done before checking
for a TypeError.