Added test for ldexp
The ruby ldexp function should be able to take a float as it's second argument.
It's just the undecorated python function that should only accept an int as
second argument because it assumes that it was already casted from float to
int.
The changes made to test_ldexp ensure that Math.ldexp can also take a float as
second argument and that it handles it right.